Add political checkbox to editions
So that we can manually change the political state of an edition add a
checkbox to the admin. This will be useful if, for example, a document
which doesn't meet the standard criteria for being political is noticed
to be political. Only managing editors and GDS staff should be able to
change the state of the political flag. The political checkbox should
only be shown after the document has been published for the first time.
This is so that we can automatically set it on first publish. If it was
editable before the first publish and editor changed it our automatic
political marking would overwrite what the editor had set.