Data migration to set Dovernment on historic Documents
The government field was added to Document in #1984, and is
set when a document is published for the first time.

This migration backfills government for Documents that have
already been published. The government of a document should
be based on the point when the document itself was first
published - distinct from when the Model was created.

This date is derived from `first_public_at` date on edition,
which reflects the historic creation date of a document.
(There is an exception for speeches, which use a different
field, but has the same intent).

From that date the government at the time can be found, and
can be set on Document.